+++ src/core/xtbinterface.cpp
@@ -35,6 +35,8 @@ XTBInterface::XTBInterface()
 
 double XTBInterface::GFN1Energy(const Molecule& molecule)
 {
+    double energy = 0;
+#ifdef USE_XTB
     int const natoms = molecule.AtomCount();
     double const charge = 0.0;
 
@@ -52,7 +54,6 @@ double XTBInterface::GFN1Energy(const Molecule& molecule)
 
     xtb::SCC_options const opt = (xtb::SCC_options){ 2, 0, 1.0, 300.0, true, true, true, 30, "none" };
 
-    double energy;
     double dipole[3];
     double q[natoms];
     double qp[6 * natoms];
@@ -60,12 +61,16 @@ double XTBInterface::GFN1Energy(const Molecule& molecule)
     char output;
     int stat = xtb::GFN1_calculation(&natoms, attyp, &charge, NULL, coord, &opt, &output,
         &energy, NULL, dipole, q, NULL);
-
+#else
+    throw("XTB is not included, sorry for that");
+#endif
     return energy;
 }
 
 double XTBInterface::GFN2Energy(const Molecule& molecule)
 {
+    double energy = 0;
+#ifdef USE_XTB
     int const natoms = molecule.AtomCount();
     double const charge = 0.0;
 
@@ -83,7 +88,6 @@ double XTBInterface::GFN2Energy(const Molecule& molecule)
 
     xtb::SCC_options const opt = (xtb::SCC_options){ 2, 1, 1.0, 300.0, true, true, true, 30, "none" };
 
-    double energy;
     double dipole[3];
     double q[natoms];
     double qp[6 * natoms];
@@ -91,15 +95,18 @@ double XTBInterface::GFN2Energy(const Molecule& molecule)
     char output;
     int stat = xtb::GFN2_calculation(&natoms, attyp, &charge, NULL, coord, &opt, &output,
         &energy, NULL, dipole, q, NULL, qp, wbo);
-
+#else
+    throw("XTB is not included, sorry for that");
+#endif
     return energy;
 }
 
 double XTBInterface::GFN2Energy(const int* attyp, const double* coord, const int natoms, const double charge, double* grad)
 {
+    double energy = 0;
+#ifdef USE_XTB
     xtb::SCC_options const opt = (xtb::SCC_options){ 2, 1, 1.0, 300.0, true, true, true, 30, "none" };
 
-    double energy;
     double dipole[3];
     double q[natoms];
     double qp[6 * natoms];
@@ -107,6 +114,8 @@ double XTBInterface::GFN2Energy(const int* attyp, const double* coord, const int
     char output;
     int stat = xtb::GFN2_calculation(&natoms, attyp, &charge, NULL, coord, &opt, &output,
         &energy, grad, dipole, q, NULL, qp, wbo);
-
+#else
+    throw("XTB is not included, sorry for that");
+#endif
     return energy;
 }