+++ Changelog.md
@@ -1,6 +1,7 @@
 # The (incomplete) curcuma Changelog
 
 ### pre Alpha
+- include optional XTB
 - prepare NEB input geometries (reordered and aligned)
 - add -led option to prepare fragment assigned xyz files (for ORCA LED calculation)
 - Add conformation filter (energy, rmsd, rank) - with automatic/forced reordering of structures